moved verifyMisbehviour function from update.go to misbehviour_handle.go
